#328 Add user on pre run works
New user is saved in the users.json file along with the other users.  The combo box is refreshed so you see the change right away. Right now the start date is set to today's date.  We need to implement a way to validate the start date if it chosen by the user in order to convert to a datetime object.