#283 Cleaned the add connection.
Refactoring the view file quite a bit.  Its now cleaner and easier to understand what is happening.  Also removed the cancel and shifted the ok button to the right.  In the controller file I only changed the names of the variables because of the changes in the view but no logic was changed.