#313, #371, I have implemented the new Geometry/Spatial View to the ModelView.
In this commit it I have renamed ViewModel and LogicModelView to ModelView and ModelCtrl respectifully. Next I applied the new spatial view to the contextview->rightclick->view details. While working on this, #371 was fixed.  The error was that it was calling the incorrect geometry function.

Plotting works for top model