#433 Cleaning the TimeSeries
This commit has some big changes. I'm moving away from the ObjectListView.py and DatabaseCtrl.py. Also the TimeSeriesData class. These files are way to big and much of it is confusing, not necessary, or the implementation could be improved. To simplify things I've recreated the TimeSeriesTab. Its in a much shorter and simpler class. The class is very dynamic and flexible. We should be consisent on the data structure however. To see the new time series tab uncomment line 37 in LowerPanelView.py and comment out line 36